+++ src/render/transform/outbound.rs
@@ -19,7 +19,7 @@ pub(super) struct Outbound;
 
 impl Transform for Outbound {
     fn enabled(&self, config: &Config) -> bool {
-        config.links.external
+        config.links.external.enabled
     }
 
     fn apply(&self, doc: &mut HtmlDocument, cx: &mut Cx<'_>) {